yearを動的に
getYearはapiからとってくる
