+++ src/usecases/getCalendar.ts
@@ -1,8 +1,5 @@
 import { Ports } from "~/adapter";
-import { SchoolCalendarEvent } from "~/api/@types";
-import { Calendar } from "~/components/PageHeader.vue";
-import { dayToJa } from "~/entities/day";
-import { EventMap } from "~/entities/event";
+import { Calendar, getSchoolCalendarEventJa } from "~/entities/calendar";
 import { getEvents } from "./getEvents";
 
 /**
@@ -15,26 +12,11 @@ export const getCalendar = async (ports: Ports): Promise<Calendar> => {
   const day = now.date();
   const week = ["日", "月", "火", "水", "木", "金", "土"][now.day()];
   const events = await getEvents(ports);
-  const schedule = getCalendarSchedule(ports, events);
+  const schedule = getSchoolCalendarEventJa(events);
   return {
     month,
     day,
     week,
     schedule,
   };
-};
-
-export const getCalendarSchedule = (
-  { dayjs }: Ports,
-  events: SchoolCalendarEvent[]
-): string => {
-  const now = dayjs();
-  const event = events.find((event) => {
-    return now.isSame(dayjs(event.date));
-  });
-  return event == null
-    ? "通常日課"
-    : event.eventType === "SubstituteDay" && event.changeTo != null
-    ? `${dayToJa(event.changeTo)}曜授業`
-    : EventMap[event.eventType];
 };
