+++ tsconfig.json
@@ -16,11 +16,5 @@
     "types": ["vite/client", "jest", "node"],
     "plugins": [{ "name": "@vuedx/typescript-plugin-vue" }]
   },
-  "include": [
-    "src/**/*.ts",
-    "src/**/*.d.ts",
-    "src/**/*.tsx",
-    "src/**/*.vue",
-    "src/components/ScheduleEditer.vue"
-  ]
+  "include": ["src/**/*.ts", "src/**/*.d.ts", "src/**/*.tsx", "src/**/*.vue"]
 }
